backend/requirements.txt
asgiref==3.7.2
attrs==25.1.0
dj-database-url==2.1.0
Django==5.0.1
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
drf-yasg==1.21.8
gunicorn==21.2.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
pillow==10.2.0
psycopg==3.2.5
psycopg-binary==3.2.5
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-dotenv==1.0.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.22.3
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
whitenoise==6.6.0
celery==5.3.1
annoy==1.17.3
scikit-learn==1.3.0
redis==4.5.5